pyproject.toml
[project]
name = "flightbrdelaypredict"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dnspython>=2.7.0",
    "mongo>=0.2.0",
    "pandas>=2.2.3",
    "ruff>=0.7.2",
    "seaborn>=0.13.2",
    "unidecode>=1.3.8",
]
